efitools: Export LD_LIBRARY_PATH in efitools recipe
Due to their build system structure, efitools do not recognize
the shared libraries residing in recipe-sysroot-native. To give
a hint to the dynmic loader, we manually set LD_LIBRARY_PATH.

If we do not set the LD_LIBRARY_PATH the native efitools binaries fail
to run due to a missing shared library as kirkstone seems to export
these paths differently.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>